% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ABC.reject}
\alias{ABC.reject}
\title{Rejection sampler for ABC estimation}
\usage{
ABC.reject(i, x0, tol, priors, Data)
}
\arguments{
\item{i}{parameter index used by \code{sapply}}

\item{x0}{List of observation vectors}

\item{tol}{Tolerance level (Euclidean distance)}

\item{priors}{List specifying parameters of the prior
distributions.}

\item{Data}{Ancillary data required by \code{foxsim}.}
}
\value{
a list with elements
\code{theta} Vector of accepted parameters
\code{xr} Vector of accepted road killed carcasses
\code{xs} Vector of accepted hunter killed carcasses
\code{pop} Vector of accepted population size for each year
}
\description{
\code{ABC.reject} is a ABC rejection sampler used by
\code{PMC.sampler}. Rejection sampling is performed for the
first specified tolerance only and samples parameters from the prior
distribution.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{ABC.weighted}}
}

