% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ABC.weighted}
\alias{ABC.weighted}
\title{Weighted rejection sampler for ABC estimation}
\usage{
ABC.weighted(i, parms, x0, tol, VarCov, w, priors, Data)
}
\arguments{
\item{i}{parameter index used by \code{sapply}}

\item{parms}{Parameter population from the previous tolerance step}

\item{x0}{List of observation vectors}

\item{tol}{Tolerance level (Euclidean distance)}

\item{VarCov}{Variance/covariance matrix for multivariate
proposals used by \code{propose.theta}.}

\item{w}{Weights matrix used to select a particular particle
from the previous population as the basis for proposal.  Used
by \code{pick.particle}.}

\item{priors}{List specifying parameters of the prior
distributions.}

\item{Data}{Ancillary data required by \code{foxsim}.}
}
\value{
a list with elements
\code{theta} Vector of accepted parameters
\code{xr} Vector of accepted road killed carcasses
\code{xs} Vector of accepted hunter killed carcasses
\code{pop} Vector of accepted population size for each year
}
\description{
\code{ABC.weighted} is a ABC rejection sampler used by
\code{PMC.sampler} that uses weighted proposals.  Weighted
sampling is used for all tolerance levels after the first.
Weights are calcuated from the previous population.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{ABC.reject}},
 \code{\link{propose.theta}},  \code{\link{pick.particle}},
 \code{\link{calc.weights}}
}

