% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc.weights}
\alias{calc.weights}
\title{Weights calculation for sequential Monte Carlo sampling}
\usage{
calc.weights(theta, thetaold, w, VarCov, priors)
}
\arguments{
\item{theta}{Current parameter population}

\item{thetaold}{Previous parameter population}

\item{w}{Vector of weights from previous population}

\item{VarCov}{Variance/covariance matrix used for multivariate
proposals.}

\item{priors}{List specifying parameters of the prior
distributions.}
}
\value{
a vector of weights for the current population
}
\description{
\code{calc.weights} calculates the weights for each particle in
the population of parameters. The weights algorithm follows Beaumont et al.
(2010) and is implemented here for multvariate normal proposals. Used in
\code{PMC.sampler}.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{propose.theta}},
  \code{\link{pick.particle}}
}

