% Generated by roxygen2 (4.0.2): do not edit by hand
\name{carcass_locs}
\alias{carcass_locs}
\title{Locations of Fox carcass discoveries in Tasmania.}
\format{A list of length 2 with the following elements\cr
xr - matrix of cell coordinates of road killed carcasses\cr
xs - matrix of cell coordinates of hunter killed carcasses\cr}
\description{
A list of length 2 representing the locations (cell row and
column numbers) of fox carcasses discovered in Tasmania
since 2001 either through road kills or hunter kills.
}

