% Generated by roxygen2 (4.0.1): do not edit by hand
\name{carcass_obs}
\alias{carcass_obs}
\title{Fox carcass discoveries in Tasmania.}
\format{A list of length 2 with the following elements\cr
xr - number of road killed carcasses each year.\cr
xs - number of hunter killed carcasses each year.\cr
Years are numbered from 2001.}
\description{
A list of length 2 representing the number of fox carcasses
discovered in Tasmania since 2001 either through road kills
or hunter kills.
}

