% Generated by roxygen2 (4.0.2): do not edit by hand
\name{display.time}
\alias{display.time}
\title{Display a pretty elapsed time}
\usage{
display.time(z)
}
\arguments{
\item{z}{Value in seconds produced by \code{difftime}}
}
\value{
Character vector indicating the elapsed time
in days, hours, minutes and seconds
}
\description{
\code{display.time} takes a \code{difftime} argument
in seconds and calculates the elapsed time
in days, hours, minutes and seconds.
}
\seealso{
\code{\link{PMC.sampler}}
}

