% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distm}
\alias{distm}
\title{Distance measure for ABC estimation}
\usage{
distm(xstar, x0)
}
\arguments{
\item{xstar}{Vector of simulated observations.}

\item{x0}{Vector of observed observations.}
}
\value{
Sum of the Euclidean distance measure
}
\description{
\code{distm} calculates the sum of Euclidean distances between observed
and simulated data. Used by \code{PMC.sampler}.
}
\seealso{
\code{\link{PMC.sampler}}
}

