% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expected.scats}
\alias{expected.scats}
\title{Expected scats calculations}
\usage{
expected.scats(pr, dr, lf)
}
\arguments{
\item{pr}{scat production rate given as No. of scats/day}

\item{dr}{scat degradation rate vector where \code{dr[1]}
is the (log) mean degradation rate per day and \code{dr[2]} is
the (log) standard deviation.}

\item{lf}{estimate of the proportion of scats that are
deposited on linear features.}
}
\value{
the (integer) number of scats
}
\description{
\code{expected.scats} calculates the number of scats that are
expected to be available for detection within the home range of
a single fox. This is calculated as the equilibrium value
between scat production and degradation.
Used by \code{scat.survey}.
}
\seealso{
\code{\link{scat.survey}}
}

