% Generated by roxygen2 (4.0.2): do not edit by hand
\name{habitatmat}
\alias{habitatmat}
\title{Fox habitat suitability map of Tasmania.}
\format{A matrix of size 114 x 106 created from a raster file
by calling \code{as.matrix(raster)}\cr}
\description{
A matrix representing fox habitat suitability on a 3 km
raster suitable for input into \code{foxsim}.  Cell values
are coded (1) Suitable habitat, (0) unsuitable habitat and
(-1) water
}

