% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has.support}
\alias{has.support}
\title{Check for prior support}
\usage{
has.support(parm, prior)
}
\arguments{
\item{parm}{Vector of parameter values}

\item{priors}{List specifying parameters of the prior
distributions.}
}
\value{
logical indicating whether all elements of \code{parm}
have non-zero prior support
}
\description{
\code{has.support} checks its arguments have non-zero prior
support.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{propose.theta}},
\code{\link{prior.density}}
}

