% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incpoints}
\alias{incpoints}
\title{List of introduction cells for foxes in Tasmania.}
\format{A list of length 2 with elements consisting of
a list of length 3 containing a matrix of cell coordinates, one
for each possible introduction point.}
\description{
A list of length 2 representing the possible introduction points
(cell coordinates) for foxes in Tasmania based on possible introductions
through the ports at Burnie, Devonport and Hobart as well as rumoured
introductions in Longford, Oatlands and St Helens.  Each element is a matrix
of cell coordinates (row and column number) of cells within 3 km of the Port
locations and 30 km of Longford, Oatlands and St Helens.
}

