% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kernel2D}
\alias{kernel2D}
\title{Dispersal kernel}
\usage{
kernel2D(eps = 3, max.disp = 30, kfun = c("circle", "Gauss", "uniform",
  "gamma"), sigma = 5, shape = 1, scale = 3)
}
\arguments{
\item{eps}{cellsize in km used for distance calculations}

\item{max.disp}{Maximum dispersal distance (km). The resulting
matrix has dimensions \code{ceiling(max.disp/eps)}.}

\item{kfun}{Character vector indicating the kernel used to calculate the cell
probabilities. Possible values are \code{"circle"}, \code{"Gauss"},
\code{"uniform"}, or \code{"gamma"}.}

\item{sigma}{Parameter for the Gaussian, circle and uniform kernels i.e. for
\code{kfun="Gauss"}, \code{kfun="circle"} or \code{kfun="uniform"}.}

\item{shape,scale}{Shape and scale parameters for the gamma distribution
\code{kfun="gamma"}.}
}
\value{
a square matrix with the dispersal probabilities for each cell
(normalised to sum to 1.0).
}
\description{
\code{Kernel2D} produces a square dispersal kernel matrix for use in
\code{foxsim}. Values are the probability of dispersal from the
origin cell (centre cell) to each cell in the matrix.
}
\examples{
kernel2D(kfun="gamma")
}

