% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Make.Raster}
\alias{Make.Raster}
\title{Combine posterior samples of spatial maps of fox occupancy saved in rds files}
\usage{
Make.Raster(X, rast, years, bw = NULL)
}
\arguments{
\item{X}{A character vector containing the relative or full paths
and filenames for all saved files.  files should be binary
files saved from \code{PMC.sampler} (by setting \code{save.post}) and
should have filenames \code{pop#.rds} where \code{#} refers to the unique
file identifer.}

\item{rast}{raster habitat file that was used for simulating fox occupancy
in \code{PMC.sampler}}

\item{years}{vector of years used for simulating fox occupancy}

\item{bw}{bandwidth (in meters) used for smoothing the final occupancy map}
}
\value{
A list of raster files of the same length as \code{years}.
Each raster in the list has the same dimensions as \code{rast}
containing the estimates of the fox occupancy probability for each cell.
}
\description{
\code{Make.Raster} combines posterior samples from the saved file(s) of
spatial maps of fox occupancy produced by \code{PMC.sampler} into
a single raster file. Cell values are the occupancy probability calculated
by averaging the presence/absence values for each cell over all the posterior
samples.  The occupancy probabilities can be optionally smoothed using
a Gaussian window with bandwidth \code{bw}.
}
\examples{
filenm<- paste0("pop",1:3,".rds") # vector of filenames
Make.Raster(filenm, habitat, 1995:2014, 3000)
}
\seealso{
\code{\link{PMC.sampler}}
}

