% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match.locations}
\alias{match.locations}
\title{Spatial Matching of cell locations}
\usage{
match.locations(x, locs, ncell, Val)
}
\arguments{
\item{x}{vector of size \code{nr} * \code{nc} representing cell
locations that will be spatially matched against the locations
in \code{locs}.}

\item{locs}{Matrix specifying the row and column number of
cell locations to match against.}

\item{ncell}{The spatial tolerance for specifying a match given
as the radial distance in cell units.}

\item{Val}{integer representing the cell value used for matching}
}
\value{
a logical vector of size \code{nr} * \code{nc} with values
set to \code{TRUE} for cell locations in \code{x} that are within
 \code{ncells} distance of locations given in \code{locs}
 (i.e. within \code{ncells} of \code{x[locs[,1] + nr * locs[,2]]}
}
\description{
\code{match.locations} performs spatial matching of cells in the vector
 \code{x} against locations given in the matrix \code{locs}. Used for
 spatial matching of simulated observations with observed locations.
 Wrapper for the C++ function \code{matchspatial}.
}
\seealso{
\code{\link{ABC.reject}}, \code{\link{ABC.weighted}}
}

