% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ModelABC}
\alias{ModelABC}
\title{Model parser for PMC}
\usage{
ModelABC(parm, Data)
}
\arguments{
\item{parm}{Vector of parameter proposals produced by \code{propose.theta}
for use in parameter estimation via \code{PMS.sampler}.}

\item{Data}{List of ancillary data required by \code{foxsim}.}
}
\value{
a list with elements
\code{xr} - simulated road killed carcasses,
\code{xs} - simulated hunter killed carcasses,
\code{pop} - population size trajectory for each year
(number of occupied cells)
}
\description{
\code{ModelABC} is a convenience function that parses parameters
and ancillary data required for \code{foxsim}. This function is used
by \code{PMC.sampler} and is not normally called by the user
}
\seealso{
\code{\link{propose.theta}}
}

