% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pick.particle}
\alias{pick.particle}
\title{Weighted particle selection}
\usage{
pick.particle(param, w)
}
\arguments{
\item{param}{Population parameter matrix}

\item{vcov}{Variance/covariance matrix for proposals}

\item{priors}{List specifying parameters of the prior
distributions.}
}
\value{
Vector of proposed parameters the same length as
\code{theta}
}
\description{
\code{pick.particle} selects a single particle (parameter
vector) from the population \code{param} at random using
probabilities based on \code{w}. Hence, particles with
higher weights have higher chance of selection.  Selected
particles are used as the basis for new proposals using
\code{propose.theta}.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{has.suport}}
}

