% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PMC.append}
\alias{PMC.append}
\title{Append PMC model objects}
\usage{
PMC.append(x, TolName)
}
\arguments{
\item{x}{A list with at least 2 elements containing \code{PMC.sampler}
objects.  Each element of \code{x} must contain samples with at least
one similar tolerance value.}

\item{TolName}{Character value indicating the tolerance value used for
appending samples.}
}
\value{
A list with elements of \code{x} appended.\cr
\code{theta} - population parameter matrix.\cr
\code{xr} - simulated road killed carcasses.\cr
\code{xs} - simulated hunter killed carcasses.\cr
\code{pop} - population size trajectory for each year
(number of occupied cells).
}
\description{
\code{PMC.append} appends posterior samples from 2 or more objects
produced by \code{PMC.sampler}.  Only posterior samples with the same
tolerance are appended and this needs to be specified using \code{TolName}.
}
\examples{
PMC.append(list(fit1, fit2), "Tol2")
}
\seealso{
\code{\link{PMC.sampler}}
}

