% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PMC.appendFiles}
\alias{PMC.appendFiles}
\title{Append PMC model objects saved in rds files}
\usage{
PMC.appendFiles(x, TolName, fileout = NULL)
}
\arguments{
\item{x}{A character vector of length at least 2 containing the relative
or full paths and filenames for all saved files.  files should be binary
files saved from \code{PMC.sampler} (by setting \code{save.post}) and all
files should contain at least one similar tolerance value.}

\item{TolName}{Character value indicating the tolerance value used for
appending samples.}

\item{fileout}{character value of the filename to write the appended object
to a file. Default is \code{NULL} which does not write anything.}
}
\value{
A list with elements of \code{x} appended.\cr
\code{theta} - population parameter matrix.\cr
\code{xr} - simulated road killed carcasses.\cr
\code{xs} - simulated hunter killed carcasses.\cr
\code{xspot} - simulated spotlight detections. \cr
\code{pop} - population size trajectory for each year
(number of occupied cells).
}
\description{
\code{PMC.appendFiles} appends posterior samples from 2 or more
\code{PMC.sampler} objects saved to rds binary files.  Only
posterior samples with the same tolerance are appended and this
needs to be specified using \code{TolName}.
}
\examples{
filenm<- paste0("post",1:3,".rds") # vector of filenames
PMC.appendFiles(filenm, "Tol2")
}
\seealso{
\code{\link{PMC.sampler}}
}

