% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pre.pad}
\alias{pre.pad}
\title{Helper function required for ABC estimation}
\usage{
pre.pad(x, obs)
}
\arguments{
\item{x}{Vector of simulated observations.}

\item{obs}{Vector of observed observations.}
}
\value{
\code{obs} with zero observations added to the beginning
of the vector so it is the same length as \code{x}.  This facilitates
distance measure calculation using \code{distm}.
}
\description{
\code{pre.pad} pads out the observation vector with zeros
so that it has the same length as the simulated observations.
Used by \code{distm}.
}
\seealso{
\code{\link{PMC.sampler}} and \code{\link{distm}}
}

