% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prior.density}
\alias{prior.density}
\title{prior density of parameter vector}
\usage{
prior.density(val, priors, logdens = FALSE)
}
\arguments{
\item{val}{Parameter values to be evaluated}

\item{priors}{List specifying parameters of the prior
distributions. The length of the list should be the same
as the number of parameters.  Each element of the list is
itself a list of length 3 indicating the name of the prior
distribution and its parameters e.g: \code{list("name", par1, par2)}.

\code{name}: Distribution name.  Currently supported distributions are \code{"uniform"}
 \code{"normal"}, \code{"lognormal"}, \code{"gamma"} and
 \code{"beta"}.\cr
 \code{par1, par2}: Parameters appropriate for the specified
 distribution.}

\item{logdens}{Logical indicating whether to calculate the log
 of the density estimate.}
}
\value{
Vector of density estimates under the prior distributions
}
\description{
\code{prior.density} evaluates the density of the supplied
parameter values in \code{val} under the prior distribution.
}
\seealso{
\code{\link{sample.priors}}, \code{\link{has.support}}
}

