% Generated by roxygen2 (4.0.1): do not edit by hand
\name{propose.theta}
\alias{propose.theta}
\title{Multivariate parameter proposals}
\usage{
propose.theta(theta, vcov, priors)
}
\arguments{
\item{theta}{Vector of parameter values to perturb}

\item{vcov}{Variance/covariance matrix for proposals}

\item{priors}{List specifying parameters of the prior
distributions.}
}
\value{
Vector of proposed parameters the same length as
\code{theta}
}
\description{
\code{propose.theta} proposes new parameter values based on
perturbing the input vector \code{theta} using a multivariate
normal distribution.  Multivariate normal proposals are
implemented using a Cholsky decomposition of \code{vcov}.
}
\seealso{
\code{\link{PMC.sampler}}, \code{\link{has.support}}
}

