% Generated by roxygen2 (4.0.2): do not edit by hand
\name{release.points}
\alias{release.points}
\title{Release points}
\usage{
release.points(intro.points, habitat, max.dist)
}
\arguments{
\item{intro.points}{Geographic coordinates of introduction points.
points must lie with the bounding box of \code{habitat}}

\item{habitat}{Raster map of habitat with suitable habitat coded as 1}

\item{max.dist}{Maximum distance (m) from \code{intro.points} to include as
a potential release cell. Only cells coded as suitable habitat (1) are
included.}
}
\value{
a list of size \code{nrow(intro.points)} with cell coordinates.
}
\description{
\code{release.points} returns cell coordinates of \code{habitat}
that are within \code{max.dist} of \code{intro.points}. Coordinates
are given as row and column number corresponding to \code{habitat}.
}
\examples{
release.points(intro, fox.suit, 10000) # suitable cells within 10km from intro.
}

