% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roadmat}
\alias{roadmat}
\title{Road network for Tasmania.}
\format{A matrix of size 114 x 106 created from a raster file
by calling \code{as.matrix(raster)}\cr}
\description{
A matrix representing the main road network for Tasmania
on a 3 km raster suitable for input into \code{foxsim}.  Cell values
are coded (1) road presence, (0) road absence
}

