% Generated by roxygen2 (4.0.1): do not edit by hand
\name{roadvec}
\alias{roadvec}
\title{Road network for Tasmania.}
\format{A vector of size 12084 created from a raster file
by calling \code{as.double(as.matrix(raster))}\cr
The dimensions of the original raster are stored in the
\code{dims} attibute (accessed by \code{attributes(roadvec)}}
\description{
A vector representing the main road network for Tasmania
on a 3 km raster suitable for input into \code{foxsim}.  Cell values
are coded (1) road presence, (0) road absence
}

