% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample.priors}
\alias{sample.priors}
\title{Sample parameters from prior distributions}
\usage{
sample.priors(priors)
}
\arguments{
\item{priors}{List specifying parameters of the prior
distributions. The length of the list should be the same
as the number of parameters.  Each element of the list is
itself a list of length 3 indicating the name of the prior
distribution and its parameters e.g: \code{list("name", par1, par2)}.

\code{name}: Distribution name.  Currently supported distributions are \code{"uniform"}
 \code{"normal"}, \code{"lognormal"}, \code{"gamma"} and
 \code{"beta"}.\cr
 \code{par1, par2}: Parameters appropriate for the specified
 distribution.}
}
\value{
Vector of random parameters from the prior distributions
}
\description{
\code{sample.priors} samples a parameter vector from the
supplied prior distribution specified in \code{priors}
}
\seealso{
\code{\link{prior.density}}, \code{\link{has.support}}
}

