% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scat.survey}
\alias{scat.survey}
\title{Scat survey monitoring observation model}
\usage{
scat.survey(occ, scatlocs, drate, parms)
}
\arguments{
\item{occ}{fox occupancy map matrix
produced by \code{PMC.sampler}.}

\item{scatlocs}{matrix containing locations of cells subject to
scat searches with the first two columns containing the cell
coordinates (row and column number) with the third column
containing the transect distance (km).}

\item{drate}{base scat detection probability (per km).}

\item{parms}{parameters for calculating the expected number
of scats available for detection (see \code{expected.scats}).}
}
\value{
matrix of the same dimension as \code{occ} with 1
indicating cells with detected scats and 0 otherwise.
}
\description{
\code{scat.survey} performs detection of foxes from scat
surveys.  Simulated scat monitoring occurs on given cells
with the probability of detection given as a exponential
function of search distance with a given base detection rate.
If the cell is occupied by a fox, then the number of scats
available for detection is given by \code{expected.scats}.
}
\seealso{
\code{\link{expected.scats}}, \code{\link{ModelABC}}
}

