% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spotlight.survey}
\alias{spotlight.survey}
\title{Spotlight monitoring observation model}
\usage{
spotlight.survey(occ, nr, spotlocs, prob, strip, cellsize)
}
\arguments{
\item{occ}{fox occupancy map vector of size \code{nr} * \code{nc}
produced by \code{PMC.sampler}.}

\item{nr}{number of rows in the map \code{occ}}

\item{spotlocs}{matrix containing cell locations of spotlight
transects with the first two columns containing the cell coordinates
(row and column number) with the third column containing the distance of
the spotlight transect in the cell (km).}

\item{prob}{base spotlight detection probability (per km).}

\item{strip}{The spatlight stripwidth}

\item{cellsize}{the size of the cells in \code{occ} (km).}
}
\value{
the total number of simulated detections
}
\description{
\code{spotlight.survey} performs detection of foxes from spotlight
surveys.  Simulated spotlight monitoring occurs on given cells
with the probability of detection simulated using a random
Poisson process with a given base detection rate.  A distance to first
detection is then simulated and compared to the transect length
in the cell to determine realised detections.
}
\seealso{
\code{\link{ModelABC}}
}

