\name{FoxSim-package}
\alias{FoxSim-package}
\docType{package}
\title{
FoxSim - Tasmanian Fox Cellular Automata
}
\description{
A cellular automata to simulate fox (Vulpes vulpes) invasion
in Tasmania coded in C++.  In addition, the package includes
functions to fit the model to observations of carcasses using
sequential Monte Carlo ABC methods.
}
\details{
\tabular{ll}{
Package: \tab FoxSim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-24\cr
License: \tab Free as in free beer\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Dave Ramsey - Arthur Rylah Institute

Maintainer: Dave Ramsey <david.ramsey@depi.vic.gov.au>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
