% Generated by roxygen2 (4.0.1): do not edit by hand
\name{incpoints}
\alias{incpoints}
\title{List of introduction cells for foxes in Tasmania.}
\format{A list of length 3 with each element consisting of
a vector of cell numbers representing possible introduction
points.}
\description{
A list of length 3 representing the possible introduction points
(cellnumbers) for foxes in Tasmania based on rumoured introductions
in Longford, Oatlands and St Helens.  Cell numbers are all habitat
cells within 20 km of each location.#'
}

